+++ src/main/java/de/otto/config/integration/helidon/spi/PropertySource.java
@@ -6,12 +6,11 @@ import lombok.extern.slf4j.Slf4j;
 import java.util.Map;
 import java.util.Set;
 
-import org.eclipse.microprofile.config.Config;
 import org.eclipse.microprofile.config.spi.ConfigProviderResolver;
 import org.eclipse.microprofile.config.spi.ConfigSource;
 
 import de.otto.config.core.Context;
-import de.otto.config.integration.helidon.config.ApplicationConfiguration;
+import de.otto.config.integration.helidon.HelidonContext;
 import de.otto.config.provider.ConfigurationProvider;
 
 @Slf4j
@@ -23,10 +22,10 @@ public class PropertySource implements ConfigSource {
     private final ConfigurationProvider configurationProvider;
 
     public PropertySource() {
-        this.context = createContext(ConfigProviderResolver.instance()
-                                                           .getBuilder()
-                                                           .addDefaultSources()
-                                                           .build());
+        this.context = HelidonContext.createContext(ConfigProviderResolver.instance()
+                                                                          .getBuilder()
+                                                                          .addDefaultSources()
+                                                                          .build());
         this.configurationProvider = ConfigurationProvider.builder()
                                                           .context(this.context)
                                                           .build();
@@ -55,11 +54,5 @@ public class PropertySource implements ConfigSource {
     @Override
     public String getValue(String propertyName) {
         return configurationProvider.getValue(propertyName);
-    }
-
-    private Context createContext(Config config) {
-        return Context.from(config.getOptionalValue("app.name", String.class).orElse("unknown"),
-                            config.getOptionalValue("mp.config.profile", String.class).orElse("local"),
-                            ApplicationConfiguration.builder().config(config).build());
     }
 }