+++ src/main/java/de/otto/config/integration/spring/env/PropertySource.java
@@ -3,7 +3,7 @@ package de.otto.config.integration.spring.env;
 import org.springframework.core.env.ConfigurableEnvironment;
 
 import de.otto.config.core.Context;
-import de.otto.config.integration.spring.config.ApplicationConfiguration;
+import de.otto.config.integration.spring.SpringContext;
 import de.otto.config.provider.ConfigurationProvider;
 import lombok.Getter;
 
@@ -16,7 +16,7 @@ public class PropertySource extends org.springframework.core.env.PropertySource<
     public PropertySource(ConfigurableEnvironment environment) {
         super(NAME);
 
-        this.context = createContext(environment);
+        this.context = SpringContext.createContext(environment);
         this.configurationProvider = ConfigurationProvider.builder()
                                                           .context(this.context)
                                                           .build();
@@ -25,11 +25,5 @@ public class PropertySource extends org.springframework.core.env.PropertySource<
     @Override
     public Object getProperty(String name) {
         return this.configurationProvider.getValue(name);
-    }
-
-    private Context createContext(ConfigurableEnvironment environment) {
-        return Context.from(environment.getProperty("spring.application.name", "unknown"),
-                            environment.getActiveProfiles().length > 0 ? environment.getActiveProfiles()[0] : "local",
-                            ApplicationConfiguration.builder().environment(environment).build());
     }
 }